LevelCompleteionProject Branch 50~60% Finished
The doors have been rearranged to add a small section of Level 1 with the new mutated rats added into it (but still flows as normal), some sections of traps and enemies added, enemy tab for Level 3, and some background wiring decoration. I'm not great with the lighting system, so the lights are somewhat lacking, and some areas are scarce in Level 3 for when we add the new trap in later. Is more or less ready to merge up. Waiting for tile colorswaps to update level visuals.