Tilemaps Added and Levels Retiled
The new tilemaps for Level 2 and 3 have been imported, sliced, and made into tilemaps for the game. They are a bit rough, and I might have done it incorrectly regarding the quality/slicing side. Levels have been retiled for anything that uses the Tile Palette. Sprites of prefabs need work (or I need help) in replacing to the proper color to match level such as the doors.